Rename Job in-memory repository source and test files
