+++ tests/job/test_inmemory_repository.py
@@ -1,4 +1,4 @@
-from middleware.job.inmemory import JobRepositoryMemory
+from middleware.job.inmemory_repository import JobRepositoryMemory
 
 
 class TestJobRepositoryMemory(object):
